feat: add theme selector
