+++ src/needs-refactoring/components/navigation/Navigation.tsx
@@ -1,15 +1,16 @@
-import { Menu, X } from "lucide-react";
 import { DropdownMenu, Flex, IconButton, Theme } from "@radix-ui/themes";
 import { t } from "@transifex/native";
+import { CheckIcon, Menu, X } from "lucide-react";
 import { useRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
 import styled from "styled-components";
+import { useTheme } from "~/hooks/useTheme";
 import { useI18nContext } from "~/modules/i18n/context/I18nContext";
 import { supportedLanguageTagsOptions } from "~/modules/i18n/i18n";
+import AppLink from "~/needs-refactoring/components/navigation/AppLink";
 import type { TranslatedAppLink } from "~/needs-refactoring/lib/useAppLink";
 import { useExpertMode } from "~/needs-refactoring/lib/useExpertMode";
 import { useNavigation } from "~/needs-refactoring/lib/useNavigation";
-import AppLink from "~/needs-refactoring/components/navigation/AppLink";
 
 function filterExpertModeLinks(
   links: TranslatedAppLink[],
@@ -56,6 +57,7 @@ export default function Navigation() {
   );
 
   const { languageLabel, setLanguageTag } = useI18nContext();
+  const { theme, setTheme } = useTheme();
 
   const dropdownMenuButton = menuLinkElements.length > 0 && (
     <Theme radius="small">
@@ -72,25 +74,44 @@ export default function Navigation() {
         </DropdownMenu.Trigger>
         <DropdownMenu.Content>
           {isExpertMode && (
-            <>
-              <DropdownMenu.Sub>
-                <DropdownMenu.SubTrigger>
-                  {languageLabel}
-                </DropdownMenu.SubTrigger>
-                <DropdownMenu.SubContent>
-                  {supportedLanguageTagsOptions.map(({ value, label }) => (
-                    <DropdownMenu.Item
-                      onClick={() => setLanguageTag(value)}
-                      key={value}
-                    >
-                      {label}
-                    </DropdownMenu.Item>
-                  ))}
-                </DropdownMenu.SubContent>
-              </DropdownMenu.Sub>
-              <DropdownMenu.Separator />
-            </>
+            <DropdownMenu.Sub>
+              <DropdownMenu.SubTrigger>{languageLabel}</DropdownMenu.SubTrigger>
+              <DropdownMenu.SubContent>
+                {supportedLanguageTagsOptions.map(({ value, label }) => (
+                  <DropdownMenu.Item
+                    onClick={() => setLanguageTag(value)}
+                    key={value}
+                  >
+                    {label}
+                  </DropdownMenu.Item>
+                ))}
+              </DropdownMenu.SubContent>
+            </DropdownMenu.Sub>
           )}
+          <DropdownMenu.Sub>
+            <DropdownMenu.SubTrigger>
+              {t("Choose theme")}
+            </DropdownMenu.SubTrigger>
+            <DropdownMenu.SubContent>
+              <DropdownMenu.Item onClick={() => setTheme("light")} key="light">
+                {t("Light")}
+                {theme === "light" && (
+                  <CheckIcon aria-hidden="true" size={16} />
+                )}
+              </DropdownMenu.Item>
+              <DropdownMenu.Item onClick={() => setTheme("dark")} key="dark">
+                {t("Dark")}
+                {theme === "dark" && <CheckIcon aria-hidden="true" size={16} />}
+              </DropdownMenu.Item>
+              <DropdownMenu.Item onClick={() => setTheme("system")} key="sytem">
+                {t("Auto")}
+                {theme === "system" && (
+                  <CheckIcon aria-hidden="true" size={16} />
+                )}
+              </DropdownMenu.Item>
+            </DropdownMenu.SubContent>
+          </DropdownMenu.Sub>
+          <DropdownMenu.Separator />
           {menuLinkElements}
         </DropdownMenu.Content>
       </DropdownMenu.Root>
